storage: delete deprecated multiIterator
Use the new NewPebbleSSTIterator() instead.

Fixes #87943

Release note: None
