+++ pkg/storage/BUILD.bazel
@@ -17,7 +17,6 @@ go_library(
         "intent_interleaving_iter.go",
         "intent_reader_writer.go",
         "min_version.go",
-        "multi_iterator.go",
         "mvcc.go",
         "mvcc_incremental_iterator.go",
         "mvcc_key.go",
@@ -114,7 +113,6 @@ go_test(
         "intent_reader_writer_test.go",
         "main_test.go",
         "min_version_test.go",
-        "multi_iterator_test.go",
         "mvcc_history_metamorphic_iterator_test.go",
         "mvcc_history_test.go",
         "mvcc_incremental_iterator_test.go",
